requirements.txt
asgiref==3.5.2
Django==4.1.3
django-mysql==4.7.1
mysql-connector-python==8.0.31
protobuf==3.20.1
sqlparse==0.4.3
tzdata==2022.6